keep old test case too
