+++ test/cases/errors/load-module-cycle-multiple/1/a.json
@@ -1,0 +1,1 @@
+"./a.json"
