+++ test/cases/errors/load-module-cycle-multiple/2/b.json
@@ -1,0 +1,1 @@
+"./a.json"
