+++ test/cases/errors/load-module-cycle-multiple/3/a.json
@@ -1,0 +1,1 @@
+"./b.json"
