+++ test/cases/errors/load-module-cycle-multiple/3/c.json
@@ -1,0 +1,1 @@
+"./a.json"
