+++ test/cases/errors/load-module-cycle-multiple/4/b.json
@@ -1,0 +1,1 @@
+"./c.json"
