+++ test/cases/errors/load-module-cycle/index.js
@@ -1,31 +1,15 @@
 it("should error loadModule when a cycle with 2 modules is requested", () => {
-	expect(require("./loader!./2/a")).toEqual([
-		["./b.json", [
-			["./a.json", "err: There is a circular build dependency, which makes it impossible to create this module"]
-		]]
-	]);
+	expect(require("./loader!./2/a")).toMatch(
+		/^source: err: There is a circular build dependency/
+	);
 });
 it("should error loadModule when a cycle with 3 modules is requested", () => {
-	expect(require("./loader!./3/a")).toEqual([
-		["./b.json", [
-			["./c.json", [
-				["./a.json", "err: There is a circular build dependency, which makes it impossible to create this module"]
-			]]
-		]]
-	]);
+	expect(require("./loader!./3/a")).toMatch(
+		/^source: source: err: There is a circular build dependency/
+	);
 });
 it("should error loadModule when requesting itself", () => {
-	expect(require("./loader!./1/a")).toEqual([
-		["./a.json", "err: There is a circular build dependency, which makes it impossible to create this module"]
-	]);
-});
-it("should not report a cycle when loadModule is used twice (https://github.com/webpack/webpack/issues/14379)", () => {
-	expect(require("./loader!./4/a")).toEqual([
-		["./b.json", [
-			["./c.json", []]
-		]],
-		["./b.json", [
-			["./c.json", []]
-		]]
-	]);
+	expect(require("./loader!./1/a")).toMatch(
+		/^err: There is a circular build dependency/
+	);
 });
