+++ test/cases/errors/load-module-cycle/loader.js
@@ -1,30 +1,12 @@
-const { promisify } = require("util");
-
 /** @type {import("../../../../").LoaderDefinitionFunction} */
 exports.default = function (source) {
-	const content = JSON.parse(source);
-	// content is one reference or an array of references
-	const refs = Array.isArray(content) ? content : [content];
+	const ref = JSON.parse(source);
 	const callback = this.async();
-	const loadModulePromise = promisify(this.loadModule.bind(this));
-
-	async function loadReferencedModules() {
-		// Modules are loaded sequentially as the false-positive circular reference
-		// bug from https://github.com/webpack/webpack/issues/14379 doesn't occur if
-		// they are loaded in parallel.
-		const loadedRefs = []
-		for(const ref of refs) {
-			try {
-				const source = await loadModulePromise("../loader!" + ref);
-				loadedRefs.push([ref, JSON.parse(source)]);
-			} catch(err) {
-				loadedRefs.push([ref, `err: ${err && err.message}`]);
-			}
+	this.loadModule("../loader!" + ref, (err, source, sourceMap, module) => {
+		if (err) {
+			callback(null, JSON.stringify(`err: ${err && err.message}`));
+		} else {
+			callback(null, JSON.stringify(`source: ${JSON.parse(source)}`));
 		}
-		return loadedRefs;
-	}
-
-	loadReferencedModules().then((loadResults) => {
-		callback(null, JSON.stringify(loadResults));
 	});
 };
